ADD: update

+++ glossy/src/sys/time.c
@@ -8,7 +8,7 @@ extern uint64_t mtime_frequency;
 
 __attribute__((weak)) unsigned int sleep(unsigned int seconds) {
   #ifdef CLINT_BASE
-    uint64_t target_tick = CLINT_getTime() + (seconds * mtime_frequency);
+    int target_tick = CLINT_getTime() + (seconds * mtime_frequency);
     while (CLINT_getTime() < target_tick) {
       asm("nop");
     }
@@ -21,12 +21,12 @@ __attribute__((weak)) unsigned int sleep(unsigned int seconds) {
 
 __attribute__((weak)) int msleep(useconds_t msec) {
   #ifdef CLINT_BASE
-  uint64_t target_tick = CLINT_getTime() + ((msec * mtime_frequency) / 1000);
-  while (CLINT_getTime() < target_tick) {
-    asm("nop");
-  }
+    int target_tick = CLINT_getTime() + ((msec * mtime_frequency) / 1000);
+    while (CLINT_getTime() < target_tick) {
+      asm("nop");
+    }
   #else
-  #warning "No CLINT peripheral found. Delay function is not available."
+    #warning "No CLINT peripheral found. Delay function is not available."
   #endif
 
   return 0;
@@ -34,12 +34,12 @@ __attribute__((weak)) int msleep(useconds_t msec) {
 
 __attribute__((weak)) int usleep(useconds_t usec) {
   #ifdef CLINT_BASE
-  uint64_t target_tick = CLINT_getTime() + ((usec * mtime_frequency) / 1000000);
-  while (CLINT_getTime() < target_tick) {
-    asm("nop");
-  }
+    int target_tick = CLINT_getTime() + ((usec * mtime_frequency) / 1000000);
+    while (CLINT_getTime() < target_tick) {
+      asm("nop");
+    }
   #else
-  #warning "No CLINT peripheral found. Delay function is not available."
+    #warning "No CLINT peripheral found. Delay function is not available."
   #endif
 
   return 0;
